Fixing failing bwc tests

+++ src/testFixtures/java/org/opensearch/neuralsearch/BaseNeuralSearchIT.java
@@ -137,10 +137,16 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
     // Wipe of all the resources after execution of the tests.
     @After
     public void cleanUp() {
-        deleteExistingIngestionPipelines();
-        deleteExistingSearchPipelines();
-        deleteExistingModels();
-        deleteExistingIndices();
+        if (shouldCleanUpResources()) {
+            deleteExistingIngestionPipelines();
+            deleteExistingSearchPipelines();
+            deleteExistingModels();
+            deleteExistingIndices();
+        }
+    }
+
+    protected boolean shouldCleanUpResources() {
+        return true;
     }
 
     protected ThreadPool setUpThreadPool() {
@@ -1614,6 +1620,33 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
                 deleteModel(m);
             }
         });
+    }
+
+    @SneakyThrows
+    protected void wipeOfTestResources(
+        final String indexName,
+        final String ingestPipeline,
+        final String modelId,
+        final String searchPipeline
+    ) {
+        if (ingestPipeline != null) {
+            deleteIngestPipeline(ingestPipeline);
+        }
+        if (searchPipeline != null) {
+            deleteSearchPipeline(searchPipeline);
+        }
+        if (modelId != null) {
+            try {
+                deleteModel(modelId);
+            } catch (AssertionError e) {
+                // sometimes we have flaky test that the model state doesn't change after call undeploy api
+                // for this case we can call undeploy api one more time
+                deleteModel(modelId);
+            }
+        }
+        if (indexName != null) {
+            deleteIndex(indexName);
+        }
     }
 
     protected float computeExpectedScore(final String modelId, final Map<String, Float> tokenWeightMap, final String queryText) {
